Renaming meadowlark-opensearch-backend to meadowlark-search-backend.
