+++ Meadowlark-js/backends/meadowlark-search-backend/docker/opensearch/docker-compose.yml
