+++ Meadowlark-js/backends/meadowlark-search-backend/src/repository/SearchException.ts
